LayoutTests/imported/w3c/web-platform-tests/content-security-policy/inheritance/history.sub-expected.txt


PASS History navigation: "about:blank" document is navigated back from history same-origin.
PASS History navigation: "about:blank" document is navigated back from history cross-origin.
FAIL History navigation: blob URL document is navigated back from history same-origin. assert_equals: Image should be blocked by CSP reloaded from history. expected "img blocked" but got "img loaded"
FAIL History navigation: blob URL document is navigated back from history cross-origin. assert_equals: Image should be blocked by CSP reloaded from history. expected "img blocked" but got "img loaded"
FAIL History navigation: blob URL document is navigated back from history (without bfcache on Firefox) same-origin. assert_equals: Image should be blocked by CSP reloaded from history. expected "img blocked" but got "img loaded"
FAIL History navigation: blob URL document is navigated back from history (without bfcache on Firefox) cross-origin. assert_equals: Image should be blocked by CSP reloaded from history. expected "img blocked" but got "img loaded"
